backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
requests
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4